[dev/core] - article init position

+++ db/db_user.py
@@ -2,6 +2,7 @@ from sqlalchemy.orm.session import Session
 from schemas import UserBase
 from db.models import DbUser
 from db import hash
+from fastapi import HTTPException, status
 
 
 def create_user(db: Session, request: UserBase):
@@ -21,11 +22,18 @@ def get_all_user(db: Session):
 
 
 def get_user(db: Session, user_id: int):
-    return db.query(DbUser).filter(DbUser.id == user_id).first()
+    user = db.query(DbUser).filter(DbUser.id == user_id).first()
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail=f"User with id {user_id} not found")
+    return user
 
 
 def update_user(db: Session, user_id: int, request: UserBase):
     user = db.query(DbUser).filter(DbUser.id == user_id)
+    if not user.first():
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail=f"User with id {user_id} not found")
     user.update({
         DbUser.username: request.username,
         DbUser.email: request.email,
@@ -37,6 +45,9 @@ def update_user(db: Session, user_id: int, request: UserBase):
 
 def delete_user(db: Session, user_id: int):
     user = db.query(DbUser).filter(DbUser.id == user_id).first()
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail=f"User with id {user_id} not found")
     db.delete(user)
     db.commit()
     return "OK"
